Module Imandra_tools

module Idf : sig ... end

module Idf_intf : sig ... end

module Region_idx : sig ... end

module Region_json : sig ... end

module Region_pp : sig ... end

module Region_pp_intf : sig ... end

module Region_probs : sig ... end

A module for creating probability distributions over regions and Boolean queries using custom hierarchical statistical models or datasets

module Region_term_synth : sig ... end

module Term_utils : sig ... end