Module type Idf_intf.SIG

A decomposition strategy

type shape =

| Tree

| DAG

| DG

type pol =

| Valid

| Invalid

| No

| Custom of string

module DSM : DSM_SIG

State machine to decompose

val pp_state : DSM.State_machine.state -> string

Printer function for a value of the state machine *

type target = DSM.Template.t list

Symbolic target of a decomposition

type node

Symbolic decomposed node of a state machine

type node_dir =

| Fwd of node

| Bwd of node

| Root of node

type path = node_dir list

Symbolic decomposed list of edges of a state machine

val node : node_dir -> node

module G : sig ... end

Decomposition graph

val paths : ?⁠from:node -> G.t -> path list

Return a list of paths from a decomposition graph starting at optional node (defaults to the root node)

val region : node -> Imandra_surface.Modular_region.t

Returns the region of a node

val expanded_constraints : node -> Imandra_surface.Term.t list

Returns the list of constraints originating from region expansion

val sample : node -> DSM.State_machine.event list

Returns the sample point of a node - a list of concrete events to get to the current node

type region_expander = string option * string list * string option

type target_step = {

event : DSM.Template.t;

valid_pol : pol;

region_expander : region_expander option;

}

val target_steps : node -> target_step list list

type replay_sample = {

state_before : DSM.State_machine.state;

event : DSM.State_machine.event;

state_after : DSM.State_machine.state;

}

sample replay output states and event

val replay_events : DSM.State_machine.event list -> replay_sample list

Replay a list of events to generate concrete state machine states from the initial state

val replay : node -> replay_sample list

Replay a node's sample events to generate concrete state machine states from the initial state

type runner

val create : ?⁠basis:string list -> ?⁠expanding:string list -> ?⁠assuming:string -> unit -> runner

val add : runner:runner -> g:G.t -> ?⁠as_:shape -> ?⁠prune_asm:bool -> ?⁠valid_pol:pol list -> ?⁠region_expander:region_expander list -> target -> unit

val run : runner -> unit -> unit

val decompose : g:G.t -> ?⁠as_:shape -> ?⁠prune_asm:bool -> ?⁠basis:string list -> ?⁠expanding:string list -> ?⁠assuming:string -> ?⁠valid_pol:pol list -> ?⁠region_expander:region_expander list -> target -> unit -> unit

Decompose a state machine using a symbolic template