Module Idf.Make

Create a new IDF module for the given SM state machine

Parameters

DSM : Idf_intf.DSM_SIG

Signature

module DSM = DSM

State machine to decompose

val pp_state : DSM.State_machine.state -> string

Printer function for a value of the state machine *

type target = DSM.Template.t list

Symbolic target of a decomposition

type node

Symbolic decomposed node of a state machine

type path = node list

Symbolic decomposed list of edges of a state machine

module G : sig ... end

Decomposition graph

val paths : ?⁠from:node -> G.t -> path list

Return a list of paths from a decomposition graph starting at optional node (defaults to the root node)

val region : node -> Imandra_surface.Modular_region.t

Returns the region of a node

val expanded_constraints : node -> Imandra_surface.Term.t list

Returns the list of constraints originating from region expansion

val sample : node -> DSM.State_machine.event list

Returns the sample point of a node - a list of concrete events to get to the current node

val sample_region : Imandra_surface.Modular_region.t -> DSM.State_machine.event list

Sample a region directly - useful if manually wanting to re-sample a region after e.g. a refinement

val target_steps : node -> (DSM.Template.t * bool option) list

The symbolic steps to get to the current node - a subset of the full target list. If target_steps = target then we are at a leaf node.

type replay_sample = {

state_before : DSM.State_machine.state;

event : DSM.State_machine.event;

state_after : DSM.State_machine.state;

}

sample replay output states and event

val replay_events : DSM.State_machine.event list -> replay_sample list

Replay a list of events to generate concrete state machine states from the initial state

val replay : node -> replay_sample list

Replay a node's sample events to generate concrete state machine states from the initial state

val decompose : g:G.t -> ?⁠basis:string list -> ?⁠include_partial_paths:bool -> ?⁠expanding:string list -> ?⁠valid_pol:bool option list -> ?⁠assuming:string -> ?⁠region_expander:(string option * string list * string option) list -> target -> unit -> unit

Decompose a state machine using a symbolic template